+++ spec/examples/add_greeting_with_hash_default.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+class AddGreetingWithHashDefault < Actor
+  input :options, default: {name: "world"}
+  output :greeting, type: String
+
+  def call
+    self.greeting = "Hello, #{options[:name]}!"
+  end
+end