+++ CHANGELOG.md
@@ -6,8 +6,9 @@ project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## unreleased
 
 Fixes:
-- missing `message` key on advanced `input` or `output` checks 
-  will be set to a default message (#117)
+- Add default message to missing `message` key on advanced `input` or `output`
+  checks (#117)
+- Allow using a Hash as `default:` (#119)
 
 ## v3.7.0